pricecompare/requirement.txt
django
python-dotenv
google-search-results
